Provide step_test_data in gerrit.get_revision_info.
This will reduce the need to manually mock steps: calls to get revision
info will now return CLs that match the requested change and patchset
numbers. Manually mocking will only be required if specific values for
other field are desired or if the test case should not have revision
info that matches the change and patchset.

Change-Id: Ibc6ec62100616680e36d9a842449ac9c6fb2440b