Fix "expr: syntax error" on macOS
The version of expr shipped on macOS does not support `substr`
command and prints `expr: syntax error`. Change the invocation
of `expr` to use the `:` operator which match rhs regexp into
lhs string and return the length of the match.

This should be supported by all versions of `expr` that are
POSIX compliant.

Bug: 1199260
Change-Id: Ib932702fedbbbc772d65086ebc371dc0a9023c0e