Update bleeding edge python versions.
The new version is built with an older Xcode version for mac-amd64,
using the 10.15 SDK, which gives us better Big Sur compatibility for
some of the very old Python libraries we have in use. In particular,
it will see the OS version as 10.16 rather than 11.0, and recognize
the 10.x-built vpython wheels as compatible.

Bug: 1175200
Change-Id: I23be64b11c8a2fb31910ce5e13b98bd5b38eeb7e