[gclient] Don't replace 'python' with 'vpython' in DEPS hooks
We don't do this for python3, and it's not clear that it's intentional
that we do this unconditionally for python2. This code as originally
added in
https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/1790603
for some other reason, with smaller scope.

After this, https://chromium-review.googlesource.com/c/chromium/src/+/2810121
actually has an effect.

DEPS files can explicitly call `vpython` when they want it, which seems
more explicit and less surprising anyways.

Bug: 1103236
Change-Id: I7418cd34a6bb8bd826a7e03467b5bed3bd8e6653