Optionally preserve line endings
CheckForWindowsLineEndings is a Chromium presubmit to make sure that
\r\n line endings (Windows line endings) don't get committed. Among
other things these line endings mess up the license checks, leading to
cryptic errors.

The problem is that ChangedContents() was stripping line endings, making
any checking for them futile. This change adds an option to preserve
line endings, to be used by Chromium's CheckForWindowsLineEndings.

It's not clear how long this has been broken.

Bug: 801033
Change-Id: I7cdb9b3581788624f9eca081da43bb070ee412a1