Fix crash during presubmit checking.
If a presubmit happened to generate a crash, the logic we have
for reporting the presubmit check times would call `six.reraise()`
incorrectly, leading to a second crash.

This CL fixes the invocation of the second crash, and also adds
an error message to help users tripping over one possible source
of the first crash (reading a binary file as text).

Bug: 1210746
Change-Id: Ic46f38901b6acf2055b3feb7272dc751dc69037c