Support Python 3.9 for Linux in depot_tools.
No Windows support right now, but we haven't built any 3.9 wheels for
Windows anyway.

We add an extra manifest file because you can't have two versions of the
same package in a CIPD manifest.

We add extra python3.{8,9}_bin_reldir.txt files for the vpython3 wrapper
script to consume, because we need them to be in alphabetical order. We
keep the existing python3_bin_reldir.txt for Windows since its scripts
are unchanged.

Bug: 1176014
Change-Id: Id779579a60af906e2dc3c728f96e86b31d8e8b48