Flush sys.stdout before reading from stdin
Some version of python by default line buffer stdout. If this
happen, and the prompt does not contain a newline, then user
won't see it until the read from stdin is complete.

Add an explicit flush of sys.stdout to ensure that even if it
is line buffered, everything will be visible by the user.

Bug: chromium:1215911
Change-Id: If93f02421b0f68954fb2c124bb04ef307dabca78