Put a call to cipd_bin_setup directly in cipd_bin_setup.sh.
This is a preliminary step in fixing metrics_python_tests dependencies.  The
next step is to add a call to cipd_bin_setup.{bat,sh} as a hook in DEPS file in
src repository to ensure all the cipd dependencies are fetched before we
isolate any tests.

Bug: 1217972
Change-Id: I8ba2345e66d10cfb03807c82039c6a7541844984
R: asvitkine@chromium.org, dpranke@google.com