[cipd] Update CIPD client to 2.6.0.
It now fetches and installs packages in parallel. This has a side
effect that the order of unzipping of packages in a manifest is
non-deterministic and depends on when particular packages finish
downloading.

To be able to show progress of parallel operations there's also
a new terminal UI with progress bars. It activates when running
CIPD at `info` logging level (default) if the stderr is attached
to a terminal (i.e. not piping to a file or the recipe engine).

Two new env vars control the new behavior:
 * CIPD_PARALLEL_DOWNLOADS: set to 0 to completely disable
   parallelism and restore pre 2.6.0 behavior. Set to 1 to fetch
   and unzip packages in paralle, but fetch at most one package
   at once. Default is 4.
 * CIPD_SIMPLE_TERMINAL_UI: set to 1 to use log-based UI instead
   of progress bars.

R=iannucci@chromium.org
BUG=1217435

Change-Id: I16815dbc397137fbb8ef09e09d0a04c77660c0c6