Harden _PresubmitResult against varying string types
When presubmit checks fail, we want to print the error message from the
underlying check/tool. We use _PresubmitResult objects to collect these
diagnostics and output them to the console and json files.

This change ensures that the string values passed to the _PresubmitResult
constructor will be converted, as needed, to 'str' instances. This way,
we will avoid encoding errors and make sure to print the actual messages
instead of repr(message).

Bug: 1225052
Change-Id: If7f5c99cb5d730c1bfbd6d108b912f77b5f2455c