Add a method to fetch the revision for a gerrit change.
eg: that's 762f1d6089690f49a20203d7a888c14830122671 for this CL:
https://chromium-review.googlesource.com/c/chromium/src/+/3162827

This will be used by code that will glue that value for a CL to gclient
to fetch. ie: CrOS bots use `gclient sync --revision src@....` to fetch
chrome. That can be used to fetch a chrome at a WIP CL, but it only
takes the CL revision, and not its ref.

Meaning, this doesn't work:
gclient sync --revision@refs/changes/27/3162827/01

But this does:
gclient sync --revision@762f1d6089690f49a20203d7a888c14830122671

So this adds a helper func to fetch that revision for a gerrit change.

Bug: 1250084
Change-Id: I3d9fa8cdeb357facf9a294cc31f58430b68bbce3