Surface yapf format errors when written to stderr
When CheckPatchFormatted presubmited canned check is executed, it runs
git cl format with --dry-run. When dry-run is used, exit codes are
ignored and we rely on only on stdout message to detect if formatting is
needed or not. This itself is wrong, but changing this will likely
require significant redesign.

This patch appends stderr output to stdout message on non-zero exit code
and therefore allowing git cl format to actually exit with code 2.

R=gavinmak@google.com

Fixed: 1264111
Change-Id: I5eccf34f533640fd93209c97d1c8d9c85bb94d83