Add a method for enforcing a tryjob.
The tryserver.require_is_tryserver method will create an infra-failing
step that raises an exception if there is no CL associated with the
build. In the case of an LED task, it will only fail so as not to cause
the removal of the builder cache in the case of a minor user error.

Change-Id: I05d3a86cc727577f067fb3eae49e46398345f672