Update to fixed vpython wheels.
The previous version was built against a different MSVC runtime than
the Python interpreter on Windows, which is unsafe on Python 2.7.

Note that this will now use a pure-Python implementation of crcmod
on Windows, since there is no packaged wheel available (only an msi,
which we don't currently support repackaging).

Bug: 1297798
Change-Id: I91e4f58cf2b994280c1620fec86dd8638d8a78ab