Handle lists of strings in fix_encoding.py
The write function in fix_encoding.py sometimes gets called with a list
of strings. This causes an exception which halts presubmits, thus hiding
all future errors. This change detects this case and changes the list of
strings to a string so that the presubmits can continue.

This was found when running "git cl presubmit --all" on Windows.

Bug: 1309977
Change-Id: I0ddacb0ca6efb2930fe23ce7c7c008e032b153f2