Handle mixed slash types in FilterSourceFile
The files_to_skip lists in chrome/android/java/src/PRESUBMIT.py were not
working on Windows because they didn't use the verbose [\\\/] construct
for path separators. Fixing this in FilterSourceFile means that multiple
failures may be fixed simultaneously. The only failures that this is
known to fix are a set of AlertDialog.Builder errors that were showing
up only on Windows.

Bug: 1309977
Change-Id: I7a10483448e00df55e2cbf8bff8bad7a5d8db124