[mac] Restore an upper limit on the number of jobs on macOS
Even when increasing the maximum number of file descriptor
as recommended in the documentation, using a large number
of concurrent jobs cause build error on certain macOS devices
(mostly iMacPro with 18 cores which ends up with a limit of
1440 jobs, and end up reaching the heightened file descriptor
limit).

Put a limit of 800 which has been found to work on most of
the devices available and still allow to have a multiplier
of 80 on highest end M1 devices (currently 10 cores).

Bug: none
Change-Id: I32560c5ae9462e94f61a773d625ef3758bf05ee5