Add Python 3 support to GetPythonUnitTests
GetPythonUnitTests (and by extension RunPythonUnitTests) always run
scripts using Python 3. This change adds a python3=False parameter to
GetPythonUnitTests to give the option of switching to Python 3. The
child scripts are run under one or the other, based on this parameter.

This change is needed in support of crrev.com/c/3679801.

Bug: 1313804
Change-Id: Ic59287352d4941707adaf7981ed7af4201b7d526