[lucicfg] Update to version 1.30.14
This is intended to release the version that includes this roll:

https://chromium.googlesource.com/infra/infra/+/839f354e1c585c7910e6e0aae2fcea1d7e321674
which includes this change:
Generate location_filters
https://chromium.googlesource.com/infra/luci/luci-go.git/+/6915bcc1e5df2847cc5b9edf2c8cac4d0cea6adc

This change ("Generate location_filters") adds a lucicfg experiment,
which, when enabled, will make lucicfg generate a new field in
commit-queue.cfg (location_filters). Existing fields are unchanged.

This CL was created following this instructions at:
https://source.chromium.org/chromium/infra/infra/+/main:go/src/go.chromium.org/luci/lucicfg/README.md#making-a-release

Bug: 1171945
Change-Id: Ifbd04ba59c5c9693c0144fbc496886becdf39d64