Add JSON formatting to git cl format
We would like to be able to format
JSON files as part of git cl format.

There are a lot of JSON files so
we will add an allowed list to limit needing
to format everything in one go.

The formatting rule is very simple, just
use python's default JSON dumps output
and add an indentation of 2 spaces.

Starting with the webview tools directory.

Bug: chromium:1335324
Change-Id: I7f3fcaff71f56fcc81459106772ec325553e3433