Make not-run tests an error again
If a presubmit test is misconfigured then it may end up running on
neither Python 2 nor Python 3. This was supposed to trigger a warning
but the warning code was broken for almost a year. It was then fixed so
that it issued an error, but that caused some breakage, so it was made a
warning.

The known problems with tests not running at all have been fixed so this
change turns the warning back into an error. This is appropriate because
a test that is not run at all is a serious bug. This change will allow
us to move more confidently when switching tests to Python 3, and it
shouldn't cause any additional breakage.

Bug: 1330859
Change-Id: I51028bb9c896c60c5cf3ccb6f472ade0bb1e0c16