Print helpful error message when updating depot_tools on Windows
The depot_tools directory may get into a state in which its no longer a
git repo, in which case we can't update it. On Windows this would cause
the update_depot_tools.bat script to fail with a somewhat confusing
error message, this CL fixes that to inform the user what happened and
what they should do about it.

Bug: 1333997
Change-Id: I2c60fc46f1fae7811ff6610e25380d08010d0597