Add an environment variable for root path.
While recursing over dependencies, there is no easy way to get the
path of the root. There are plenty of less easy way (such as having a
script that runs before the recursion. Or removing the relative path
to pwd).

If this patch is accepted, I plan to use it in order to create a
script that require both root path and dependency path.

Change-Id: I21fa37d05078ecbe59185cc12c229d40576308a2