Treat UnitTests failures as presubmit --all errors
When upload presubmits are being run then GetUnitTests would treat
errors as warnings, so as to not stop developers from uploading their
changes. This makes sense when patches are being manually uploaded, but
doesn't make sense when running a "git cl presubmit --all --upload" bot.
This behavior caused nine failures (but not all failures) to be missed
when doing these tests.

This change makes these failures errors when testing with --all
--upload, but leaves them as warnings for other --upload purposes.

Bug: 1309977
Change-Id: Ibf149475e4cdee10bbbbc86fd0ab668b7a679089